# [vectorized] [join] eliminate branch prediction & opt hash join performance