--- a/be/src/vec/exec/join/vhash_join_node.cpp
+++ b/be/src/vec/exec/join/vhash_join_node.cpp
@@ -42,13 +42,14 @@ using ProfileCounter = RuntimeProfile::Counter;
 template <class HashTableContext, bool ignore_null, bool build_unique>
 struct ProcessHashTableBuild {
     ProcessHashTableBuild(int rows, Block& acquired_block, ColumnRawPtrs& build_raw_ptrs,
-                          HashJoinNode* join_node, int batch_size)
+                          HashJoinNode* join_node, int batch_size, uint8_t blk_ind)
             : _rows(rows),
               _skip_rows(0),
               _acquired_block(acquired_block),
               _build_raw_ptrs(build_raw_ptrs),
               _join_node(join_node),
-              _batch_size(batch_size) {}
+              _batch_size(batch_size),
+              _blk_ind(blk_ind) {}
 
     Status operator()(HashTableContext& hash_table_ctx, ConstNullMapPtr null_map,
                       bool has_runtime_filter) {
@@ -88,28 +89,20 @@ struct ProcessHashTableBuild {
             }
 
             if (emplace_result.is_inserted()) {
-                new (&emplace_result.get_mapped()) Mapped({&_acquired_block, k});
+                new (&emplace_result.get_mapped()) Mapped({_blk_ind, k});
                 if (has_runtime_filter) {
                     inserted_rows.push_back(k);
                 }
             } else {
                 if constexpr (!build_unique) {
                     /// The first element of the list is stored in the value of the hash table, the rest in the pool.
-                    emplace_result.get_mapped().insert({&_acquired_block, k}, _join_node->_arena);
+                    emplace_result.get_mapped().insert({_blk_ind, k}, _join_node->_arena);
                     if (has_runtime_filter) {
                         inserted_rows.push_back(k);
                     }
                 } else {
                     _skip_rows++;
                 }
-            }
-        }
-
-        if constexpr (build_unique) {
-            // If all row in build block is skip, just remove it
-            // to reduce mem pressure
-            if (_skip_rows == _rows) {
-                _join_node->_acquire_list.remove_last_element();
             }
         }
 
@@ -126,6 +119,7 @@ private:
     ColumnRawPtrs& _build_raw_ptrs;
     HashJoinNode* _join_node;
     int _batch_size;
+    uint8_t _blk_ind;
 };
 
 template <class HashTableContext>
@@ -168,6 +162,7 @@ struct ProcessHashTableProbe {
               _right_col_len(join_node->_right_col_len),
               _batch_size(batch_size),
               _probe_rows(probe_rows),
+              _build_block(join_node->_build_block),
               _probe_block(join_node->_probe_block),
               _probe_index(join_node->_probe_index),
               _probe_raw_ptrs(join_node->_probe_columns),
@@ -187,6 +182,8 @@ struct ProcessHashTableProbe {
         std::vector<uint32_t> items_counts(_probe_rows);
         auto& mcol = mutable_block.mutable_columns();
         int current_offset = 0;
+        std::vector<std::pair<uint8_t, uint32_t>> _build_index;
+        _build_index.reserve(1.2 * _batch_size);
 
         for (; _probe_index < _probe_rows;) {
             if constexpr (ignore_null) {
@@ -196,7 +193,6 @@ struct ProcessHashTableProbe {
                 }
             }
             int repeat_count = 0;
-
             auto find_result = (*null_map)[_probe_index]
                                 ? decltype(key_getter.find_key(hash_table_ctx.hash_table, _probe_index,
                                                                 _arena)) {nullptr, false}
@@ -208,22 +204,14 @@ struct ProcessHashTableProbe {
                     // We should rethink whether to use this iterator mode in the future. Now just opt the one row case
                     if (mapped.get_row_count() == 1) {
                         ++repeat_count;
-                        for (size_t j = 0; j < _right_col_len; ++j) {
-                            auto& column = *mapped.block->get_by_position(j).column;
-                            mcol[j + _right_col_idx]->insert_from(column, mapped.row_num);
-                        }
+                        _build_index.emplace_back(std::make_pair(mapped.block_index, mapped.row_num));
                     } else {
                         // prefetch is more useful while matching to multiple rows
                         if (_probe_index + 2 < _probe_rows)
                             key_getter.prefetch(hash_table_ctx.hash_table, _probe_index + 2, _arena);
                         for (auto it = mapped.begin(); it.ok(); ++it) {
                             ++repeat_count;
-                            for (size_t j = 0; j < _right_col_len; ++j) {
-                                auto& column = *it->block->get_by_position(j).column;
-                                // TODO: interface insert from cause serious performance problems
-                                //  when column is nullable. Try to make more effective way
-                                mcol[j + _right_col_idx]->insert_from(column, it->row_num);
-                            }
+                            _build_index.emplace_back(std::make_pair(it->block_index, it->row_num));
                         }
                     }
                 }
@@ -250,10 +238,7 @@ struct ProcessHashTableProbe {
                         if constexpr (JoinOpType::value != TJoinOp::RIGHT_ANTI_JOIN && 
                                       JoinOpType::value != TJoinOp::RIGHT_SEMI_JOIN) {
                             ++repeat_count;
-                            for (size_t j = 0; j < _right_col_len; ++j) {
-                                auto& column = *mapped.block->get_by_position(j).column;
-                                mcol[j + _right_col_idx]->insert_from(column, mapped.row_num);
-                            }
+                            _build_index.emplace_back(std::make_pair(mapped.block_index, mapped.row_num));
                         }
                     } else {
                         for (auto it = mapped.begin(); it.ok(); ++it) {
@@ -262,12 +247,7 @@ struct ProcessHashTableProbe {
                             if constexpr (JoinOpType::value != TJoinOp::RIGHT_ANTI_JOIN && 
                                           JoinOpType::value != TJoinOp::RIGHT_SEMI_JOIN) {
                                 ++repeat_count;
-                                for (size_t j = 0; j < _right_col_len; ++j) {
-                                    auto& column = *it->block->get_by_position(j).column;
-                                    // TODO: interface insert from cause serious performance problems
-                                    //  when column is nullable. Try to make more effective way
-                                    mcol[j + _right_col_idx]->insert_from(column, it->row_num);
-                                }
+                                _build_index.emplace_back(std::make_pair(it->block_index, it->row_num));
                             }
                             it->visited = true;
                         }
@@ -284,12 +264,28 @@ struct ProcessHashTableProbe {
             }
             items_counts[_probe_index++] = repeat_count;
             current_offset += repeat_count;
-
             if (current_offset >= _batch_size) {
                 break;
             }
         }
-        
+
+        // insert all match build rows
+        if (_build_block.size() == 1) {
+            for (int i = 0; i < _right_col_len; i++) {
+                    auto &column = *_build_block[0].get_by_position(i).column;
+                for (int j = 0; j < _build_index.size(); j++) {
+                    mcol[i + _right_col_idx]->insert_from(column, _build_index[j].second);
+                }
+            }
+        } else {
+            for (int i = 0; i < _right_col_len; i++) {
+                for (int j = 0; j < _build_index.size(); j++) {
+                    auto &column = *_build_block[_build_index[j].first].get_by_position(i).column;
+                    mcol[i + _right_col_idx]->insert_from(column, _build_index[j].second);
+                }
+            }
+        }
+
         for (int i = 0; i < _right_col_idx; ++i) {
             auto& column = _probe_block.get_by_position(i).column;
             column->replicate(items_counts.data(), current_offset, *mcol[i]);
@@ -322,6 +318,9 @@ struct ProcessHashTableProbe {
         std::vector<bool> same_to_prev;
         same_to_prev.reserve(1.2 * _batch_size);
 
+        std::vector<std::pair<uint8_t, uint32_t>> _build_index;
+        _build_index.reserve(1.2 * _batch_size);
+
         int current_offset = 0;
 
         for (; _probe_index < _probe_rows;) {
@@ -344,10 +343,7 @@ struct ProcessHashTableProbe {
 
                 for (auto it = mapped.begin(); it.ok(); ++it) {
                     ++current_offset;
-                    for (size_t j = 0; j < _right_col_len; ++j) {
-                        auto& column = *it->block->get_by_position(j).column;
-                        mcol[j + _right_col_idx]->insert_from(column, it->row_num);
-                    }
+                    _build_index.emplace_back(std::make_pair(it->block_index, it->row_num));
                     visited_map.emplace_back(&it->visited);
                 }
                 same_to_prev.emplace_back(false);
@@ -385,6 +381,14 @@ struct ProcessHashTableProbe {
 
         for (int i = _probe_index; i < _probe_rows; ++i) {
             offset_data[i] = current_offset;
+        }
+
+        // insert all match build rows
+        for (int i = 0; i < _right_col_len; i++) {
+            for (int j = 0; j < _build_index.size(); j++) {
+                auto &column = *_build_block[_build_index[j].first].get_by_position(i).column;
+                mcol[i + _right_col_idx]->insert_from(column, _build_index[j].second);
+            }
         }
         output_block->swap(mutable_block.to_block());
         for (int i = 0; i < _right_col_idx; ++i) {
@@ -507,10 +511,10 @@ struct ProcessHashTableProbe {
         auto& iter = hash_table_ctx.iter;
         auto block_size = 0;
 
-        auto insert_from_hash_table = [&](const Block* block, uint32_t row_num) {
+        auto insert_from_hash_table = [&](uint8_t block_index, uint32_t row_num) {
             block_size++;
             for (size_t j = 0; j < _right_col_len; ++j) {
-                auto& column = *block->get_by_position(j).column;
+                auto& column = *_build_block[block_index].get_by_position(j).column;
                 mcol[j + _right_col_idx]->insert_from(column, row_num);
             }
         };
@@ -520,10 +524,10 @@ struct ProcessHashTableProbe {
             for (auto it = mapped.begin(); it.ok(); ++it) {
                 if constexpr (JoinOpType::value == TJoinOp::RIGHT_SEMI_JOIN) {
                     if (it->visited)
-                        insert_from_hash_table(it->block, it->row_num);
+                        insert_from_hash_table(it->block_index, it->row_num);
                 } else {
                     if (!it->visited)
-                        insert_from_hash_table(it->block, it->row_num);
+                        insert_from_hash_table(it->block_index, it->row_num);
                 }
             }
         }
@@ -551,6 +555,7 @@ private:
     const int _right_col_len;
     const int _batch_size;
     const size_t _probe_rows;
+    const std::vector<Block>& _build_block;
     const Block& _probe_block;
     int& _probe_index;
     ColumnRawPtrs& _probe_raw_ptrs;
@@ -862,7 +867,12 @@ Status HashJoinNode::_hash_table_build(RuntimeState* state) {
     RETURN_IF_ERROR(child(1)->open(state));
     SCOPED_TIMER(_build_timer);
     Block block;
-
+    int right_col_len = _right_table_data_types.size();
+    MutableColumns columns(right_col_len);
+    for (int i = 0; i < right_col_len; i++) {
+        columns[i] = _right_table_data_types[i]->create_column();
+    }
+    uint8_t index = 0;
     bool eos = false;
     while (!eos) {
         block.clear();
@@ -873,9 +883,34 @@ Status HashJoinNode::_hash_table_build(RuntimeState* state) {
         _mem_used += block.allocated_bytes();
         RETURN_IF_LIMIT_EXCEEDED(state, "Hash join, while getting next from the child 1.");
 
-        RETURN_IF_ERROR(_process_build_block(state, block));
-        RETURN_IF_LIMIT_EXCEEDED(state, "Hash join, while constructing the hash table.");
+        if (block.rows() != 0) {
+            int i = 0;
+            for (const auto &data : block) {
+                const auto & column = *data.column.get();
+                columns[i].get()->insert_range_from(column, 0, block.rows());
+                i++;
+            }
+        }
+
+        if (_mem_used > 2 * 1024UL * 1024UL * 1024UL) {
+            for (int i = 0; i < right_col_len; ++i) {
+                _build_block.emplace_back(Block({ColumnWithTypeAndName(std::move(columns[i]),_right_table_data_types[i], "")}));
+            }
+            RETURN_IF_ERROR(_process_build_block(state, _build_block[index], index));
+            RETURN_IF_LIMIT_EXCEEDED(state, "Hash join, while constructing the hash table.");
+            ++index;
+            for (int i = 0; i < right_col_len; i++) {
+                columns[i] = _right_table_data_types[i]->create_column();
+            }
+        }
     }
+
+    for (int i = 0; i < _right_table_data_types.size(); ++i) {
+        _build_block.emplace_back(Block({ColumnWithTypeAndName(std::move(columns[i]),_right_table_data_types[i], "")}));
+    }
+
+    RETURN_IF_ERROR(_process_build_block(state, _build_block[index], index));
+    RETURN_IF_LIMIT_EXCEEDED(state, "Hash join, while constructing the hash table.");
 
     return std::visit(
             [&](auto&& arg) -> Status {
@@ -975,7 +1010,7 @@ Status HashJoinNode::extract_probe_join_column(Block& block, NullMap& null_map,
     return Status::OK();
 }
 
-Status HashJoinNode::_process_build_block(RuntimeState* state, Block& block) {
+Status HashJoinNode::_process_build_block(RuntimeState* state, Block& block, uint8_t blk_ind) {
     SCOPED_TIMER(_build_table_timer);
     size_t rows = block.rows();
     if (rows == 0) {
@@ -983,8 +1018,7 @@ Status HashJoinNode::_process_build_block(RuntimeState* state, Block& block) {
     }
     COUNTER_UPDATE(_build_rows_counter, rows);
 
-    auto& acquired_block = _acquire_list.acquire(std::move(block));
-    materialize_block_inplace(acquired_block);
+    materialize_block_inplace(block);
 
     ColumnRawPtrs raw_ptrs(_build_expr_ctxs.size());
 
@@ -997,7 +1031,7 @@ Status HashJoinNode::_process_build_block(RuntimeState* state, Block& block) {
             [&](auto&& arg) -> Status {
                 using HashTableCtxType = std::decay_t<decltype(arg)>;
                 if constexpr (!std::is_same_v<HashTableCtxType, std::monostate>) {
-                    return extract_build_join_column(acquired_block, null_map_val, raw_ptrs,
+                    return extract_build_join_column(block, null_map_val, raw_ptrs,
                                                      has_null, *_build_expr_call_timer);
                 } else {
                     LOG(FATAL) << "FATAL: uninited hash table";
@@ -1014,7 +1048,7 @@ Status HashJoinNode::_process_build_block(RuntimeState* state, Block& block) {
                 if constexpr (!std::is_same_v<HashTableCtxType, std::monostate>) {
 #define CALL_BUILD_FUNCTION(HAS_NULL, BUILD_UNIQUE)                                           \
     ProcessHashTableBuild<HashTableCtxType, HAS_NULL, BUILD_UNIQUE> hash_table_build_process( \
-            rows, acquired_block, raw_ptrs, this, state->batch_size());                       \
+            rows, block, raw_ptrs, this, state->batch_size(), blk_ind);                       \
     st = hash_table_build_process(arg, &null_map_val, has_runtime_filter);
                     if (std::pair {has_null, _build_unique} == std::pair {true, true}) {
                         CALL_BUILD_FUNCTION(true, true);