… be/src/vec/exec/vset_operation_node.cpp
… be/src/vec/exec/vset_operation_node.h