# Proposed changes

Issue Number: close #xxx

## Problem Summary:

Opt hash join performance

1. Eliminate hashjoin probe phase branch perdiction by make_bool_variant function in compile time。
especially for innerjoin (most scenes)， left semi join left anti join(mostly used).
2. Reduce mapped memory usage，the hash map traverse mapped is time consuming，so delete the block pointer（8 Bytes）is cache friendly.

After opt, the performance is greatly improved。


My test:(SSB benchmark)
set runime_filter_type=0;
set parallel_fragment_exec_instance_num = 1;
1. inner join: 
(before opt)
MySQL [ssb1]> SELECT count(c_custkey) FROM lineorder,customer WHERE lo_custkey = c_custkey;
+--------------------+
| count(`c_custkey`) |
+--------------------+
|          600037902 |
+--------------------+
1 row in set (43.25 sec)
(after opt)
MySQL [ssb1]> SELECT count(c_custkey) FROM lineorder,customer WHERE lo_custkey = c_custkey;
+--------------------+
| count(`c_custkey`) |
+--------------------+
|          600037902 |
+--------------------+
1 row in set (30.25 sec)

2. ssb q3.3
(before opt)
MySQL [ssb1]> SELECT c_city, s_city, d_year, SUM(lo_revenue) AS  REVENUE FROM customer, lineorder, supplier, dates WHERE lo_custkey = c_custkey AND lo_suppkey = s_suppkey AND  lo_orderdate = d_datekey AND  (c_city='UNITED KI1' OR c_city='UNITED KI5') AND (s_city='UNITED KI1' OR s_city='UNITED KI5') AND d_year >= 1992 AND d_year <= 1997 GROUP BY c_city, s_city, d_year ORDER BY d_year ASC,  REVENUE DESC;
+------------+------------+--------+------------+
| c_city     | s_city     | d_year | REVENUE    |
+------------+------------+--------+------------+
| UNITED KI1 | UNITED KI1 |   1992 | 5776096629 |
| UNITED KI5 | UNITED KI1 |   1992 | 5555883901 |
| UNITED KI5 | UNITED KI5 |   1992 | 5348705805 |
| UNITED KI1 | UNITED KI5 |   1992 | 5326870427 |
| UNITED KI1 | UNITED KI1 |   1993 | 5892974670 |
| UNITED KI1 | UNITED KI5 |   1993 | 5490859451 |
| UNITED KI5 | UNITED KI1 |   1993 | 5468354303 |
| UNITED KI5 | UNITED KI5 |   1993 | 5089909647 |
| UNITED KI5 | UNITED KI1 |   1994 | 5437315108 |
| UNITED KI1 | UNITED KI1 |   1994 | 5348775917 |
| UNITED KI5 | UNITED KI5 |   1994 | 5310936695 |
| UNITED KI1 | UNITED KI5 |   1994 | 5237461110 |
| UNITED KI1 | UNITED KI1 |   1995 | 5737551920 |
| UNITED KI5 | UNITED KI5 |   1995 | 5657584590 |
| UNITED KI5 | UNITED KI1 |   1995 | 5260093556 |
| UNITED KI1 | UNITED KI5 |   1995 | 5213763257 |
| UNITED KI5 | UNITED KI1 |   1996 | 5522325005 |
| UNITED KI1 | UNITED KI1 |   1996 | 5451244409 |
| UNITED KI5 | UNITED KI5 |   1996 | 5231759057 |
| UNITED KI1 | UNITED KI5 |   1996 | 5203962897 |
| UNITED KI1 | UNITED KI1 |   1997 | 5340760807 |
| UNITED KI5 | UNITED KI1 |   1997 | 5295685214 |
| UNITED KI1 | UNITED KI5 |   1997 | 5188428156 |
| UNITED KI5 | UNITED KI5 |   1997 | 5024634475 |
+------------+------------+--------+------------+
24 rows in set (32.15 sec)
(after opt): 
MySQL [ssb1]> SELECT c_city, s_city, d_year, SUM(lo_revenue) AS  REVENUE FROM customer, lineorder, supplier, dates WHERE lo_custkey = c_custkey AND lo_suppkey = s_suppkey AND  lo_orderdate = d_datekey AND  (c_city='UNITED KI1' OR c_city='UNITED KI5') AND (s_city='UNITED KI1' OR s_city='UNITED KI5') AND d_year >= 1992 AND d_year <= 1997 GROUP BY c_city, s_city, d_year ORDER BY d_year ASC,  REVENUE DESC;
+------------+------------+--------+------------+
| c_city     | s_city     | d_year | REVENUE    |
+------------+------------+--------+------------+
| UNITED KI1 | UNITED KI1 |   1992 | 5776096629 |
| UNITED KI5 | UNITED KI1 |   1992 | 5555883901 |
| UNITED KI5 | UNITED KI5 |   1992 | 5348705805 |
| UNITED KI1 | UNITED KI5 |   1992 | 5326870427 |
| UNITED KI1 | UNITED KI1 |   1993 | 5892974670 |
| UNITED KI1 | UNITED KI5 |   1993 | 5490859451 |
| UNITED KI5 | UNITED KI1 |   1993 | 5468354303 |
| UNITED KI5 | UNITED KI5 |   1993 | 5089909647 |
| UNITED KI5 | UNITED KI1 |   1994 | 5437315108 |
| UNITED KI1 | UNITED KI1 |   1994 | 5348775917 |
| UNITED KI5 | UNITED KI5 |   1994 | 5310936695 |
| UNITED KI1 | UNITED KI5 |   1994 | 5237461110 |
| UNITED KI1 | UNITED KI1 |   1995 | 5737551920 |
| UNITED KI5 | UNITED KI5 |   1995 | 5657584590 |
| UNITED KI5 | UNITED KI1 |   1995 | 5260093556 |
| UNITED KI1 | UNITED KI5 |   1995 | 5213763257 |
| UNITED KI5 | UNITED KI1 |   1996 | 5522325005 |
| UNITED KI1 | UNITED KI1 |   1996 | 5451244409 |
| UNITED KI5 | UNITED KI5 |   1996 | 5231759057 |
| UNITED KI1 | UNITED KI5 |   1996 | 5203962897 |
| UNITED KI1 | UNITED KI1 |   1997 | 5340760807 |
| UNITED KI5 | UNITED KI1 |   1997 | 5295685214 |
| UNITED KI1 | UNITED KI5 |   1997 | 5188428156 |
| UNITED KI5 | UNITED KI5 |   1997 | 5024634475 |
+------------+------------+--------+------------+
24 rows in set (11.95 sec)
----------------------------------------------------

## Checklist(Required)

1. Does it affect the original behavior: (No)
2. Has unit tests been added: (No Need)
3. Has document been added or modified: (No Need)
4. Does it need to update dependencies: (No)
5. Are there any changes that cannot be rolled back: (No)

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...